Der KD-Tree vergisst nun nicht mehr seine Teil KD-Trees. Die Abfrage muss nun noch bearbeitet werden.

+++ KDTree.h
@@ -5,7 +5,7 @@ class KDTree
 {
 public:
 	KDTree(std::vector<Point3d>& points, int dim = 0);
-	std::vector<Point3d> abfrage(int laenge, Point3d& point, int dim);
+	std::vector<Point3d> abfrage(double laenge, Point3d& point, int dim);
 	Point3d value;
 	KDTree* left;
 	KDTree* right;
